feat: SSR for iconMap in icon loader
