+++ src/routes/+layout.server.ts
@@ -1,9 +1,11 @@
 import type { LayoutServerLoad } from "./$types";
 import { getGitCommitHash, getGitHTTPUrl } from "$lib/utils/gitUtils";
+import iconMap from "$lib/utils/iconMap";
 
 export const load: LayoutServerLoad = async () => {
 	return {
         gitCommitHash: getGitCommitHash(),
-		gitHTTPUrl: getGitHTTPUrl()
+		gitHTTPUrl: getGitHTTPUrl(),
+		iconMap
 	};
 };
